% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchkdeleter-package.R
\docType{package}
\name{rchkdeleter-package}
\alias{rchkdeleter}
\alias{rchkdeleter-package}
\title{rchkdeleter: Try to Reproduce an 'rchk' Deleter Bug}
\description{
Tries to reproduce a bug with 'rchk'.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net}

}
\keyword{internal}
